Version Control Assignment
Cloned repository and forked off a branch. Renamed repository to "Data-Glacier-Internship-Fall-2023" and created a branch of that repository called "Data-Glacier-Internship-Fall-2023". For the assignment, I ran the code "add.py" script in Python and added my name and favorite sport to the script. This updated the "response.json" file in the repository. I then made a "Week 1" folder and added the response json file to that folder for week 1. Then I pulled up the repository in Anaconda and ran the test code to run the script. There were no errors and it went successfully. This push is for the assignment.